requirements.txt
requests~=2.32.3
beautifulsoup4~=4.13.3
python-dotenv~=1.0.1
rich~=13.9.4